linera-core: keep round synchronization in the extracted sender, opt-in per consumer
